refactor: clubTime -> club
