+++ dodam-application/dodam-rest-api/src/main/java/b1nd/dodam/restapi/club/application/ClubMemberUseCase.java
@@ -3,7 +3,7 @@ package b1nd.dodam.restapi.club.application;
 import b1nd.dodam.domain.rds.club.enumeration.ClubStatus;
 import b1nd.dodam.domain.rds.club.enumeration.ClubTimeType;
 import b1nd.dodam.domain.rds.club.service.ClubMemberService;
-import b1nd.dodam.domain.rds.club.service.ClubTimeService;
+import b1nd.dodam.domain.rds.club.service.ClubService;
 import b1nd.dodam.domain.rds.member.entity.Member;
 import b1nd.dodam.domain.rds.member.entity.Student;
 import b1nd.dodam.domain.rds.member.repository.StudentRepository;
@@ -24,13 +24,13 @@ import java.util.List;
 @RequiredArgsConstructor
 @Transactional(rollbackFor = Exception.class)
 public class ClubMemberUseCase {
+    private final ClubService clubService;
     private final ClubMemberService clubMemberService;
-    private final ClubTimeService clubTimeService;
     private final MemberAuthenticationHolder authenticationHolder;
     private final StudentRepository studentRepository;
 
     public Response joinClubs(List<JoinClubMemberReq> reqs) {
-        clubTimeService.validateApplicationDuration(ClubTimeType.CLUB_APPLICANT);
+        clubService.validateApplicationDuration(ClubTimeType.CLUB_APPLICANT);
         Student student = studentRepository.getByMember(authenticationHolder.current());
         boolean clubJoined = clubMemberService.isCreativeClubJoined(student);
         if (clubJoined) {
