+++ dodam-application/dodam-rest-api/src/main/java/b1nd/dodam/restapi/club/application/ClubUseCase.java
@@ -1,19 +1,21 @@
 package b1nd.dodam.restapi.club.application;
 
 import b1nd.dodam.domain.rds.club.entity.Club;
+import b1nd.dodam.domain.rds.club.entity.ClubTime;
 import b1nd.dodam.domain.rds.club.enumeration.ClubStatus;
 import b1nd.dodam.domain.rds.club.enumeration.ClubTimeType;
 import b1nd.dodam.domain.rds.club.service.ClubMemberService;
 import b1nd.dodam.domain.rds.club.service.ClubService;
-import b1nd.dodam.domain.rds.club.service.ClubTimeService;
 import b1nd.dodam.domain.rds.member.entity.Student;
 import b1nd.dodam.domain.rds.member.repository.StudentRepository;
 import b1nd.dodam.domain.rds.member.repository.TeacherRepository;
 import b1nd.dodam.restapi.auth.infrastructure.security.support.MemberAuthenticationHolder;
+import b1nd.dodam.restapi.club.application.data.req.ClubTimeReq;
 import b1nd.dodam.restapi.club.application.data.req.CreateClubReq;
 import b1nd.dodam.restapi.club.application.data.req.UpdateClubInfoReq;
 import b1nd.dodam.restapi.club.application.data.req.UpdateClubReq;
 import b1nd.dodam.restapi.club.application.data.res.ClubDetailRes;
+import b1nd.dodam.restapi.club.application.data.res.ClubTimeRes;
 import b1nd.dodam.restapi.support.data.Response;
 import b1nd.dodam.restapi.support.data.ResponseData;
 import lombok.RequiredArgsConstructor;
@@ -27,14 +29,18 @@ import java.util.List;
 @Transactional(rollbackFor = Exception.class)
 public class ClubUseCase {
     private final ClubService clubService;
-    private final ClubTimeService clubTimeService;
     private final ClubMemberService clubMemberService;
     private final StudentRepository studentRepository;
     private final MemberAuthenticationHolder authHolder;
     private final TeacherRepository teacherRepository;
 
+    public Response save(ClubTimeReq req) {
+        clubService.setClubTime(req.toEntity());
+        return Response.created("시간 설정 성공");
+    }
+
     public Response save(CreateClubReq req) {
-        clubTimeService.validateApplicationDuration(ClubTimeType.CLUB_CREATED);
+        clubService.validateApplicationDuration(ClubTimeType.CLUB_CREATED);
         clubService.checkIsNameDuplicated(req.name());
         Club club = req.toEntity();
         Student leader = studentRepository.getByMember(authHolder.current());
@@ -71,6 +77,12 @@ public class ClubUseCase {
         clubs.forEach(c -> c.updateStatus(req.status(), req.reason()));
         clubService.updateAll(clubs);
         return Response.ok("동아리 상태 변경 성공");
+    }
+
+    public ResponseData<ClubTimeRes> find() {
+        ClubTime createTime = clubService.getClubTime(ClubTimeType.CLUB_CREATED);
+        ClubTime applicantTime = clubService.getClubTime(ClubTimeType.CLUB_APPLICANT);
+        return ResponseData.ok("시간 불러오기 성공", ClubTimeRes.of(createTime, applicantTime));
     }
 
     public Response updateInfo(Long id, UpdateClubInfoReq req) {
