+++ dodam-system-domain/dodam-domain-rds/src/main/java/b1nd/dodam/domain/rds/club/service/ClubService.java
@@ -2,15 +2,20 @@ package b1nd.dodam.domain.rds.club.service;
 
 import b1nd.dodam.domain.rds.club.entity.Club;
 import b1nd.dodam.domain.rds.club.entity.ClubMember;
+import b1nd.dodam.domain.rds.club.entity.ClubTime;
 import b1nd.dodam.domain.rds.club.enumeration.ClubPermission;
 import b1nd.dodam.domain.rds.club.enumeration.ClubStatus;
+import b1nd.dodam.domain.rds.club.enumeration.ClubTimeType;
+import b1nd.dodam.domain.rds.club.exception.ClubApplicationDurationPassedException;
 import b1nd.dodam.domain.rds.club.exception.ClubDuplicateException;
 import b1nd.dodam.domain.rds.club.repository.ClubMemberRepository;
 import b1nd.dodam.domain.rds.club.repository.ClubRepository;
+import b1nd.dodam.domain.rds.club.repository.ClubTimeRepository;
 import b1nd.dodam.domain.rds.member.entity.Student;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -19,6 +24,7 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class ClubService {
     private final ClubRepository clubRepository;
+    private final ClubTimeRepository clubTimeRepository;
     private final ClubMemberRepository clubMemberRepository;
     private final String DELETED_PREFIX =  "_deleted";
 
@@ -59,6 +65,22 @@ public class ClubService {
     public void deleteClub(Club club) {
         club.updateStatus(club.getName() + DELETED_PREFIX, ClubStatus.DELETED);
         clubRepository.save(club);
+    }
+
+    public void validateApplicationDuration(ClubTimeType clubTimeType) {
+        ClubTime time = getClubTime(clubTimeType);
+        LocalDate today = LocalDate.now();
+        if (today.isBefore(time.getStart()) || today.isAfter(time.getEnd())) {
+            throw new ClubApplicationDurationPassedException();
+        }
+    }
+
+    public ClubTime getClubTime(ClubTimeType clubTimeType) {
+        return clubTimeRepository.findById(clubTimeType).orElseThrow(ClubApplicationDurationPassedException::new);
+    }
+
+    public void setClubTime(ClubTime clubTime) {
+        clubTimeRepository.save(clubTime);
     }
 
     private ClubMember createMember(Club club, Student student, ClubPermission clubPermission, ClubStatus clubStatus) {
