--- dodam-system-domain/dodam-domain-rds/src/main/java/b1nd/dodam/domain/rds/club/service/ClubTimeService.java
@@ -1,32 +1,0 @@
-package b1nd.dodam.domain.rds.club.service;
-
-import b1nd.dodam.domain.rds.club.entity.ClubTime;
-import b1nd.dodam.domain.rds.club.enumeration.ClubTimeType;
-import b1nd.dodam.domain.rds.club.exception.ClubApplicationDurationPassedException;
-import b1nd.dodam.domain.rds.club.repository.ClubTimeRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.time.LocalDate;
-
-@Service
-@RequiredArgsConstructor
-public class ClubTimeService {
-    private final ClubTimeRepository clubTimeRepository;
-
-    public ClubTime getClubTime(ClubTimeType clubTimeType) {
-        return clubTimeRepository.findById(clubTimeType).orElseThrow(ClubApplicationDurationPassedException::new);
-    }
-
-    public void setClubTime(ClubTime clubTime) {
-        clubTimeRepository.save(clubTime);
-    }
-
-    public void validateApplicationDuration(ClubTimeType clubTimeType) {
-        ClubTime time = getClubTime(clubTimeType);
-        LocalDate today = LocalDate.now();
-        if (today.isBefore(time.getStart()) || today.isAfter(time.getEnd())) {
-            throw new ClubApplicationDurationPassedException();
-        }
-    }
-}
